Extract test setup from build_validation_package to actions/mod
